Update: Player Input + Camera
Camera moves more or less how I'd want it to.

The player movement is kinda broken but it's more progress somewhat. Need to be able to make the player look in the direction they are moving next. Movement doesn't take into account where the camera is looking either, this needs to be fixed.

[CommitChecker]