Update: States
Implemented basic state machine structure. Added StateMachine node to the player scene. Need to work on converting locomotion logic from player.cpp -> state_locomotion (with the ability to input necessary values...). StateMachine has access to the current_state and the current state is able to let the state machine know what state to transition to next (using handle_transition) - there are the standard godot functions available for states to use as well as a draw_debug for imgui purposes.

Would be nice to be able to hide "New State" from the resource drop down in the inspector but I don't know how to do that yet and it's not massively important right now.