Update: Attack State
Added in a new attack animation with a SFX.

Updated the godot-cpp submodule to include keeping the MAKE_TYPED_ARRAY macros included, this allows GDExtensions to make arrays of custom types and display them in editor (see StateMachine::states).

Updated the build commands to store a scons cache for godot so it shouldn't take as long. I noticed it was having to do a full recompile when switching between working on something in visual studio, then using the commit checker to see if my work compiles correctly for different platforms/configurations.

[CommitChecker]