Update: Development Configuration
Added a new configuration which allows me to use a more iterative workflow. I can run a locally built godot binary file and then switch to development configuration in visual studio and just build the GDExtension code - this allows me to hot reload my game code changes with the editor running and saves on build times slightly.

Also reduced the resolution of the initial window size and changed it from fullscreen to windowed.