Update: Fix Player Camera Controls
Updated the locomotion to use the global rotation y value of the camera class as the rotation value for calculating direction with. This means the player should always move in the direction the camera is facing. Added extra imgui debug to help with debugging this if it's still a bug in builds.

Added new input for imgui focus. This allows me to bring up the imgui debug menus, select the objects I want to see the debug information for, and then unfocus imgui to have full control in game whilst displaying any debug I need to check. Can toggle this using Shift + D on keyboard or by pressing in the Left Stick and pressing DPad Down on a joypad.

Also updated the display window in builds to be maximized so they should fit the user's screen regardless of resolution, hopefully...