Update: Virtual Joystick
Implemented virtual joystick class and added them as children of the player UI node. They will only appear if the platform is Android or iOS. Left virtual joystick will move the character. Right virtual joystick will move the camera around. These joysticks function in the same way as the PC/controller controls do and feel roughly similar in how they handle input.

Added new textures for the virtual joysticks.

Fixed compile issue (on Android/Web/iOS) in build_information where some imgui logic was not inside of the #if check.

Updated the build script so that development for other platforms can be iterated on too. E.g. if you want to work on a gameplay feature and test on android, you can have the editor running as a binary file, change your configuration in visual studio to development and select android as the platform. Compile it. Then use the one click deploy feature to see your changes running on your device.

Fixed issue with scons script where it wouldn't prefix lib for android binary files created on windows. This was causing a disparity between what was being output and what the gdextension file expected.

[CommitChecker]